refactor: extract UI components and styles to meet file size limits
- ExerciseCard: extract modals → ExerciseCardModals, helpers/styles → ExerciseCardHelpers
- SetInputRow: extract ReadOnlyCells + styles → SetInputHelpers
- AddExerciseModal: extract ExerciseRow, use shared ChipSelect for muscle groups
- CreateExerciseModal: replace inline chip sections with shared ChipSelect
- QuickExerciseModal: extract styles → QuickExerciseModalStyles
- WorkoutScreen: extract styles → WorkoutScreenStyles
- New shared ChipSelect component for reusable chip-selector groups

All files now under 250 lines per CONTRIBUTING.md limits.
